fix: arrumando estrutura dos dtos
