+++ src/main/java/org/example/apifluxar/controller/BatchController.java
@@ -3,9 +3,9 @@ package org.example.apifluxar.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import org.example.apifluxar.dto.BatchRequestDTO;
-import org.example.apifluxar.dto.BatchResponseCreateDTO;
-import org.example.apifluxar.dto.BatchResponseDTO;
+import org.example.apifluxar.dto.batch.BatchRequestDTO;
+import org.example.apifluxar.dto.batch.BatchResponseCreateDTO;
+import org.example.apifluxar.dto.batch.BatchResponseDTO;
 import org.example.apifluxar.service.BatchService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
