+++ src/main/java/org/example/apifluxar/controller/EmployeeController.java
@@ -4,9 +4,9 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.groups.Default;
-import org.example.apifluxar.dto.EmployeeResponseDTO;
-import org.example.apifluxar.dto.EmployeeRequestDTO;
-import org.example.apifluxar.dto.UpdatePhotoRequestDTO;
+import org.example.apifluxar.dto.employee.EmployeeResponseDTO;
+import org.example.apifluxar.dto.employee.EmployeeRequestDTO;
+import org.example.apifluxar.dto.employee.UpdatePhotoRequestDTO;
 import org.example.apifluxar.service.EmployeeService;
 import org.example.apifluxar.validation.OnCreate;
 import org.example.apifluxar.validation.OnPatch;
