+++ src/main/java/org/example/apifluxar/controller/ProductController.java
@@ -3,10 +3,9 @@ package org.example.apifluxar.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import org.example.apifluxar.dto.AllProductsResponseDTO;
-import org.example.apifluxar.dto.ProductRequestDTO;
-import org.example.apifluxar.dto.ProductResponseDTO;
-import org.example.apifluxar.model.Product;
+import org.example.apifluxar.dto.products.AllProductsResponseDTO;
+import org.example.apifluxar.dto.products.ProductRequestDTO;
+import org.example.apifluxar.dto.products.ProductResponseDTO;
 import org.example.apifluxar.service.ProductService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
