+++ src/main/java/org/example/apifluxar/dto/batch/BatchResponseDTO.java
@@ -1,8 +1,11 @@
-package org.example.apifluxar.dto;
+package org.example.apifluxar.dto.batch;
+
+import org.example.apifluxar.dto.unit.UnitResponseDTO;
+import org.example.apifluxar.dto.products.ProductResponseDTO;
 
 import java.time.LocalDate;
 
-public class BatchResponseCreateDTO {
+public class BatchResponseDTO {
     private String idLote;
     private LocalDate validade;
     private Double altura;
@@ -12,7 +15,7 @@ public class BatchResponseCreateDTO {
     private ProductResponseDTO product;
     private UnitResponseDTO unit;
 
-    public BatchResponseCreateDTO(Double altura, Double comprimento, String idLote, Double largura, ProductResponseDTO product, UnitResponseDTO unit, LocalDate validade) {
+    public BatchResponseDTO(Double altura, Double comprimento, String idLote, Double largura, ProductResponseDTO product, UnitResponseDTO unit, LocalDate validade) {
         this.altura = altura;
         this.comprimento = comprimento;
         this.idLote = idLote;
@@ -22,7 +25,7 @@ public class BatchResponseCreateDTO {
         this.validade = validade;
     }
 
-    public BatchResponseCreateDTO(String idLote, LocalDate validade, Double altura, Double comprimento, Double largura, Double volume) {
+    public BatchResponseDTO(String idLote, LocalDate validade, Double altura, Double comprimento, Double largura, Double volume) {
         this.idLote = idLote;
         this.validade = validade;
         this.altura = altura;
@@ -59,6 +62,14 @@ public class BatchResponseCreateDTO {
         return comprimento;
     }
 
+    public Double getVolume() {
+        return volume;
+    }
+
+    public void setVolume(Double volume) {
+        this.volume = volume;
+    }
+
     public void setComprimento(Double comprimento) {
         this.comprimento = comprimento;
     }
@@ -71,12 +82,14 @@ public class BatchResponseCreateDTO {
         this.largura = largura;
     }
 
-    public Double getVolume() {
-        return volume;
+
+
+    public ProductResponseDTO getProduct() {
+        return product;
     }
 
-    public void setVolume(Double volume) {
-        this.volume = volume;
+    public void setProduct(ProductResponseDTO product) {
+        this.product = product;
     }
 
     public UnitResponseDTO getUnit() {
@@ -85,13 +98,5 @@ public class BatchResponseCreateDTO {
 
     public void setUnit(UnitResponseDTO unit) {
         this.unit = unit;
-    }
-
-    public ProductResponseDTO getProduct() {
-        return product;
-    }
-
-    public void setProduct(ProductResponseDTO product) {
-        this.product = product;
     }
 }
