+++ src/main/java/org/example/apifluxar/dto/industry/IndustryResponseDTO.java
@@ -1,4 +1,4 @@
-package org.example.apifluxar.dto;
+package org.example.apifluxar.dto.industry;
 
 public class IndustryResponseDTO {
     private String nome;
