+++ src/main/java/org/example/apifluxar/service/CapacityHistoryService.java
@@ -1,7 +1,10 @@
 package org.example.apifluxar.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.example.apifluxar.dto.*;
+import org.example.apifluxar.dto.capacityHistory.CapacityHistoryResponseDTO;
+import org.example.apifluxar.dto.industry.IndustryResponseDTO;
+import org.example.apifluxar.dto.sector.SectorResponseDTO;
+import org.example.apifluxar.dto.unit.UnitResponseDTO;
 import org.example.apifluxar.model.*;
 import org.example.apifluxar.repository.CapacityHistoryRepository;
 import org.springframework.http.HttpStatus;
@@ -46,7 +49,7 @@ public class CapacityHistoryService {
                     unit.getEstado(),
                     unit.getNumero(),
                     unit.getBairro(),
-                    objectMapper.convertValue(unit.getIndustry(),IndustryResponseDTO.class)
+                    objectMapper.convertValue(unit.getIndustry(), IndustryResponseDTO.class)
 
             );
             unitDTO.setId(unit.getId());
