+++ src/main/java/org/example/apifluxar/service/StockHistoryService.java
@@ -1,19 +1,12 @@
 package org.example.apifluxar.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.example.apifluxar.dto.BatchResponseDTO;
-import org.example.apifluxar.dto.StockHistoryResponseDTO;
-import org.example.apifluxar.model.Batch;
+import org.example.apifluxar.dto.stockHistory.StockHistoryResponseDTO;
 import org.example.apifluxar.model.StockHistory;
 import org.example.apifluxar.repository.StockHistoryRepository;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
 
 
 @Service
