+++ src/main/java/org/example/apifluxar/service/UnitService.java
@@ -1,10 +1,8 @@
 package org.example.apifluxar.service;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.example.apifluxar.dto.CapacityHistoryResponseDTO;
-import org.example.apifluxar.dto.IndustryResponseDTO;
+import org.example.apifluxar.dto.industry.IndustryResponseDTO;
 import org.example.apifluxar.model.*;
-import org.example.apifluxar.dto.UnitResponseDTO;
+import org.example.apifluxar.dto.unit.UnitResponseDTO;
 import org.example.apifluxar.repository.UnitRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
